Create unit tests for local
The testcases for local are written such that they temporarily
overide functions such that we don't have to mess with creating
files for testing. Since Julia is trying to optimize compilation of
methods we can end up in situations where we overwrite a method
used inside of another method and it continues to use the original
method. Using `[]...` allows us to change the behaviour of a
function.

See https://github.com/JuliaLang/julia/issues/265 for details on the
core issue.